Create new PictureUploadUrl and PictureUrl functions to get SAS tokens for read/write of profile pics to blob storage.

+++ GC-API/FunctionsTests/Helpers/DocumentDBRepository.cs
@@ -167,7 +167,8 @@ namespace FunctionsTests.Helpers
 
         private static async Task DeleteDatabaseAsync()
         {
-            await Client.DeleteDatabaseAsync(UriFactory.CreateDatabaseUri(DatabaseId));
+            if (Client != null)
+                await Client.DeleteDatabaseAsync(UriFactory.CreateDatabaseUri(DatabaseId));
         }
     }
 }
